Disabled the flawed marked content parsing in PDF Clown
This bug is discussed on stack overflow:

PDFClown Different font-size in one line
https://stackoverflow.com/questions/45413682/pdfclown-different-font-size-in-one-line

The issue is caused by a bug in PDF Clown: it assumes that marked content sections and save/restore graphics state blocks are properly contained in each other and don't overlap.

The only easy way I found to repair this was to disable the marked content parsing in PDF Clown.